Make the tests work again
