+++ tests/bootstrap.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once __DIR__ . '/../wp-includes/sqlite/class-wp-sqlite-query-rewriter.php';
+require_once __DIR__ . '/../wp-includes/sqlite/class-wp-sqlite-lexer.php';
+require_once __DIR__ . '/../wp-includes/sqlite/class-wp-sqlite-token.php';
+require_once __DIR__ . '/../wp-includes/sqlite/class-wp-sqlite-pdo-user-defined-functions.php';
+require_once __DIR__ . '/../wp-includes/sqlite/class-wp-sqlite-translator.php';
